+++ src/writers/progress_file_reader.go
@@ -46,7 +46,7 @@ func (pr *ProgressFileReader) Read(p []byte) (int, error) {
 	if n > 0 {
 		newRead := atomic.AddInt64(&pr.read, int64(n))
 		if pr.size > 0 {
-			percentage := int(float64(newRead) * 100 / float64(pr.size))
+			percentage := float64(newRead) * 100 / float64(pr.size)
 			if pr.ns != nil {
 				message := pr.prefix
 				if message == "" {
@@ -78,7 +78,7 @@ func (pr *ProgressFileReader) ReadAt(p []byte, off int64) (int, error) {
 		if newRead > pr.size {
 			newRead = pr.size
 		}
-		percentage := int(float64(newRead) * 100 / float64(pr.size))
+		percentage := float64(newRead) * 100 / float64(pr.size)
 		if pr.ns != nil {
 			message := pr.prefix
 			if message == "" {
